+++ firebase-vscode/src/data-connect/execution/execution.ts
@@ -10,45 +10,38 @@ import {
   ExecutionItem,
   ExecutionState,
   createExecution,
-  executionArgsJSON,
   selectExecutionId,
   selectedExecution,
   selectedExecutionId,
   updateExecution,
 } from "./execution-store";
-import { batch, effect, Signal } from "@preact/signals-core";
+import { batch, effect } from "@preact/signals-core";
 import {
   OperationDefinitionNode,
   OperationTypeNode,
   print,
   buildClientSchema,
   validate,
-  DocumentNode,
-  Kind,
-  TypeNode,
   parse,
 } from "graphql";
 import { DataConnectService } from "../service";
 import { DataConnectError, toSerializedError } from "../../../common/error";
-import { OperationLocation } from "../types";
 import { InstanceType } from "../code-lens-provider";
 import { DATA_CONNECT_EVENT_NAME, AnalyticsLogger } from "../../analytics";
-import { getDefaultScalarValue } from "../ad-hoc-mutations";
 import { EmulatorsController } from "../../core/emulators";
 import { getConnectorGQLText, insertQueryAt } from "../file-utils";
 import { pluginLogger } from "../../logger-wrapper";
 import * as gif from "../../../../src/gemini/fdcExperience";
 import { ensureGIFApiTos } from "../../../../src/dataconnect/ensureApis";
 import { configstore } from "../../../../src/configstore";
+import { executionAuthParams, executionVarsJSON, ExecutionParamsService } from "./execution-params";
+import { ExecuteGraphqlRequest } from "../../dataconnect/types";
 
-interface TypedInput {
-  varName: string;
-  type: string | null;
-}
-
-interface ExecutionInput {
-  ast: OperationDefinitionNode;
-  location: OperationLocation;
+export interface ExecutionInput {
+  operationAst: OperationDefinitionNode;
+  document: string;
+  documentPath: string;
+  position: vscode.Position;
   instance: InstanceType;
 }
 
@@ -60,12 +53,11 @@ export interface GenerateOperationInput {
   existingQuery: string;
 }
 
-export const lastExecutionInputSignal = new Signal<ExecutionInput | null>(null);
-
 export function registerExecution(
   context: ExtensionContext,
   broker: ExtensionBrokerImpl,
   dataConnectService: DataConnectService,
+  paramsService: ExecutionParamsService,
   analyticsLogger: AnalyticsLogger,
   emulatorsController: EmulatorsController,
 ): Disposable {
@@ -88,13 +80,14 @@ export function registerExecution(
 
   function notifyDataConnectResults(item: ExecutionItem) {
     broker.send("notifyDataConnectResults", {
-      args: item.args ?? "{}",
-      query: print(item.operation),
+      displayName: `${item.input.operationAst.operation} ${item.input.operationAst.name?.value ?? ""}`,
+      query: print(item.input.operationAst),
       results:
         item.results instanceof Error
           ? toSerializedError(item.results)
           : item.results,
-      displayName: item.operation.operation,
+      variables: item.variables || "",
+      auth: item.auth,
     });
   }
 
@@ -115,21 +108,14 @@ export function registerExecution(
 
   // re run called from execution panel;
   const rerunExecutionBroker = broker.on("rerunExecution", () => {
-    if (!lastExecutionInputSignal.value) {
-      return;
+    const item = selectedExecution.value;
+    if (item) {
+      executeOperation(item.input);
     }
-    executeOperation(
-      lastExecutionInputSignal.value.ast,
-      lastExecutionInputSignal.value.location,
-      lastExecutionInputSignal.value.instance,
-    );
   });
 
-  async function executeOperation(
-    ast: OperationDefinitionNode,
-    { document, documentPath, position }: OperationLocation,
-    instance: InstanceType,
-  ) {
+  async function executeOperation(arg: ExecutionInput) {
+    const { operationAst: ast, document, documentPath, instance } = arg;
     analyticsLogger.logger.logUsage(
       instance === InstanceType.LOCAL
         ? DATA_CONNECT_EVENT_NAME.RUN_LOCAL
@@ -142,17 +128,9 @@ export function registerExecution(
     );
     await vscode.window.activeTextEditor?.document.save();
 
-    // hold last execution in memory, and send operation name to webview
-    lastExecutionInputSignal.value = {
-      ast,
-      location: { document, documentPath, position },
-      instance,
-    };
-    broker.send("notifyLastOperation", ast.name?.value ?? "anonymous");
-
     // focus on execution panel immediately
     vscode.commands.executeCommand(
-      "data-connect-execution-configuration.focus",
+      "data-connect-execution-parameters.focus",
     );
 
     const configs = vscode.workspace.getConfiguration("firebase.dataConnect");
@@ -249,92 +227,51 @@ export function registerExecution(
         return;
       }
     }
-    
 
-    // if execution args is empty, reset to {}
-    if (!executionArgsJSON.value) {
-      executionArgsJSON.value = "{}";
+    const servicePath = await dataConnectService.servicePath(documentPath);
+    if (!servicePath) {
+      throw new Error("No service found for document path: " + documentPath);
     }
-
-    // Check for missing arguments
-    const missingArgs = await verifyMissingArgs(ast, executionArgsJSON.value);
-
-    // prompt user to continue execution or modify arguments
-    if (missingArgs.length > 0) {
-      analyticsLogger.logger.logUsage(DATA_CONNECT_EVENT_NAME.MISSING_VARIABLES);
-      // open a modal with option to run anyway or edit args
-      const editArgs = { title: "Edit variables" };
-      const continueExecution = { title: "Continue Execution" };
-      const result = await vscode.window.showInformationMessage(
-        `Missing required variables. Would you like to modify them?`,
-        { modal: !process.env.VSCODE_TEST_MODE },
-        editArgs,
-        continueExecution,
-      );
-
-      if (result === editArgs) {
-        const missingArgsJSON = getDefaultArgs(missingArgs);
-
-        // combine w/ existing args, and send to webview
-        const newArgsJsonString = JSON.stringify({
-          ...JSON.parse(executionArgsJSON.value),
-          ...missingArgsJSON,
-        });
-
-        broker.send("notifyDataConnectArgs", newArgsJsonString);
-        return;
-      }
-    }
+    const req: ExecuteGraphqlRequest = {
+      operationName: ast.name?.value,
+      variables: paramsService.executeGraphqlVariables(),
+      query: gqlText || document,
+      extensions: paramsService.executeGraphqlExtensions(),
+    };
 
     const item = createExecution({
       label: ast.name?.value ?? "anonymous",
       timestamp: Date.now(),
       state: ExecutionState.RUNNING,
-      operation: ast,
-      args: executionArgsJSON.value,
-      documentPath,
-      position,
+      input: arg,
+      variables: executionVarsJSON.value,
+      auth: executionAuthParams.value,
+      results: new Error("missing results"),
     });
-
-    function updateAndSelect(updates: Partial<ExecutionItem>) {
-      batch(() => {
-        updateExecution(item.executionId, { ...item, ...updates });
-        selectExecutionId(item.executionId);
-      });
-    }
 
     try {
       // Execute queries/mutations from their source code.
       // That ensures that we can execute queries in unsaved files.
-
-      const results = await dataConnectService.executeGraphQL({
-        operationName: ast.name?.value,
-        // We send the compiled GQL from the whole connector to support fragments
-        // In the case of adhoc operation, just send the sole document
-        query: gqlText || document,
-        variables: executionArgsJSON.value,
-        path: documentPath,
-        instance,
-      });
-
-      updateAndSelect({
-        state:
-          // Executing queries may return a response which contains errors
-          // without throwing.
-          // In that case, we mark the execution as errored.
-          (results.errors?.length ?? 0) > 0
-            ? ExecutionState.ERRORED
-            : ExecutionState.FINISHED,
-        results,
-      });
+      const results = await dataConnectService.executeGraphQL(servicePath, instance, req);
+      // Executing queries may return a response which contains errors
+      item.state = (results.errors?.length ?? 0) > 0
+        ? ExecutionState.ERRORED
+        : ExecutionState.FINISHED;
+      item.results = results;
     } catch (error) {
-      updateAndSelect({
-        state: ExecutionState.ERRORED,
-        results:
-          error instanceof Error
-            ? error
-            : new DataConnectError("Unknown error", error),
-      });
+      item.state = ExecutionState.ERRORED;
+      item.results = error instanceof Error
+        ? error
+        : new DataConnectError("Unknown error", error);
+    }
+
+    batch(() => {
+      updateExecution(item.executionId, item);
+      selectExecutionId(item.executionId);
+    });
+
+    if (item.state === ExecutionState.ERRORED) {
+      await paramsService.applyDetectedFixes(ast);
     }
   }
 
@@ -397,19 +334,13 @@ ${schema}
     return false;
   }
 
-  const sub4 = broker.on(
-    "definedDataConnectArgs",
-    (value) => (executionArgsJSON.value = value),
-  );
-
   return Disposable.from(
     { dispose: sub1 },
     { dispose: sub2 },
     { dispose: sub3 },
-    { dispose: sub4 },
     { dispose: rerunExecutionBroker },
     registerWebview({
-      name: "data-connect-execution-configuration",
+      name: "data-connect-execution-parameters",
       context,
       broker,
     }),
@@ -421,8 +352,8 @@ ${schema}
     executionHistoryTreeView,
     vscode.commands.registerCommand(
       "firebase.dataConnect.executeOperation",
-      async (ast, location, instanceType: InstanceType) => {
-        await executeOperation(ast, location, instanceType);
+      async (arg: ExecutionInput) => {
+        await executeOperation(arg);
       },
     ),
     vscode.commands.registerCommand(
@@ -451,68 +382,4 @@ function executionError(message: string, error?: string) {
     `Failed to execute operation: ${message}: \n${JSON.stringify(error, undefined, 2)}`,
   );
   throw new Error(error);
-}
-
-function getArgsWithTypeFromOperation(
-  ast: OperationDefinitionNode,
-): TypedInput[] {
-  if (!ast.variableDefinitions) {
-    return [];
-  }
-  return ast.variableDefinitions.map((variable) => {
-    const varName = variable.variable.name.value;
-
-    const typeNode = variable.type;
-
-    function getType(typeNode: TypeNode): string | null {
-      // Same as previous example
-      switch (typeNode.kind) {
-        case "NamedType":
-          return typeNode.name.value;
-        case "ListType":
-          const innerTypeName = getType(typeNode.type);
-          return `[${innerTypeName}]`;
-        case "NonNullType":
-          const nonNullTypeName = getType(typeNode.type);
-          return `${nonNullTypeName}!`;
-        default:
-          return null;
-      }
-    }
-
-    const type = getType(typeNode);
-
-    return { varName, type };
-  });
-}
-
-// checks if required arguments are present in payload
-async function verifyMissingArgs(
-  ast: OperationDefinitionNode,
-  jsonArgs: string,
-): Promise<TypedInput[]> {
-  let userArgs: { [key: string]: any };
-  try {
-    userArgs = JSON.parse(jsonArgs);
-  } catch (e: any) {
-    executionError("Invalid JSON: ", e);
-    return [];
-  }
-
-  const argsWithType = getArgsWithTypeFromOperation(ast);
-  if (!argsWithType) {
-    return [];
-  }
-  return argsWithType
-    .filter((arg) => arg.type?.includes("!"))
-    .filter((arg) => !userArgs[arg.varName]);
-}
-
-function getDefaultArgs(args: TypedInput[]) {
-  return args.reduce((acc: { [key: string]: any }, arg) => {
-    const defaultValue = getDefaultScalarValue(arg.type as string);
-
-    acc[arg.varName] = defaultValue;
-    return acc;
-  }, {});
 }