+++ src/dataconnect/provisionCloudSql.ts
@@ -8,6 +8,7 @@ import { checkFreeTrialInstanceUsed, freeTrialTermsLink } from "./freeTrial";
 import { promiseWithSpinner } from "../utils";
 import { trackGA4 } from "../track";
 import * as utils from "../utils";
+import { Source } from "../init/features/dataconnect";
 
 const GOOGLE_ML_INTEGRATION_ROLE = "roles/aiplatform.user";
 
@@ -24,7 +25,7 @@ export async function setupCloudSql(args: {
   instanceId: string;
   databaseId: string;
   requireGoogleMlIntegration: boolean;
-  source: "init" | "mcp_init" | "deploy";
+  source: Source;
   dryRun?: boolean;
 }): Promise<void> {
   const { projectId, instanceId, requireGoogleMlIntegration, dryRun } = args;
@@ -37,7 +38,7 @@ export async function setupCloudSql(args: {
     success = true;
   } finally {
     if (!dryRun) {
-      await trackGA4(
+      void trackGA4(
         "dataconnect_cloud_sql",
         {
           source: args.source,
@@ -76,9 +77,10 @@ async function upsertInstance(
       `Found existing Cloud SQL instance ${clc.bold(instanceId)}.`,
     );
     stats.databaseVersion = existingInstance.databaseVersion;
-    stats.dataconnectLabel = existingInstance.settings?.userLabels?.["firebase-data-connect"] as
-      | cloudSqlAdminClient.DataConnectLabel
-      | undefined;
+    stats.dataconnectLabel =
+      (existingInstance.settings?.userLabels?.[
+        "firebase-data-connect"
+      ] as cloudSqlAdminClient.DataConnectLabel) || "absent";
 
     const why = getUpdateReason(existingInstance, requireGoogleMlIntegration);
     if (why) {